Added text search filter for strings
