+++ client/package.json
@@ -19,6 +19,7 @@
         "axios": "^1.9.0",
         "react": "^19.0.0",
         "react-dom": "^19.0.0",
+        "react-highlight-words": "^0.21.0",
         "react-router-dom": "^7.6.0"
     },
     "devDependencies": {
