+++ client/src/pages/products/Edit/components/AttributesField.tsx
